appplatform: splitting and renaming the spring cloud service
