+++ azurerm/internal/services/appplatform/validate/spring_cloud_service_name_test.go
